[Fix] Changed from reference type to integer type

+++ db/migrate/20230618192134_create_relationships.rb
@@ -1,9 +1,8 @@
 class CreateRelationships < ActiveRecord::Migration[6.1]
   def change
     create_table :relationships do |t|
-      t.bigint  :answer_id, null: false, foreign_key: true
+      t.integer :answer_id, null: false, foreign_key: true
       t.integer :tag_id,    null: false, foreign_key: true
-
       t.timestamps
     end
   end
